GAN/requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.2
dm-tree==0.1.5
gast==0.2.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-probability==0.8.0
tensorflow-graphics-gpu==1.0.0
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
pandas==1.2.3
scikit-learn==0.24.1
scikit-image==0.18.1